feat(ui): move the busy indicator into the agent message
- jumping dots + rotating verb render as the last line of a streaming
  agent message, hidden while queued or awaiting approval; removed
  from the input bar
- clamp long approval titles (6 lines) and ellipsize tool-row titles
